Finish splitting of ReceiptPage
