+++ src/components/index.ts
@@ -1,5 +1,4 @@
 export { CompactDeletableListItem, ListColor } from './CompactDeletableListItem';
-export { SummaryPage } from '../features/apiDelegation/components/SummaryPage';
 export { DeletableListItem } from './DeletableListItem';
 export { Line } from './Line';
 export { PageContainer } from './PageContainer';
