+++ src/features/apiDelegation/components/ErrorAlert/ErrorAlert.module.css
@@ -3,3 +3,7 @@
   display: flex;
   justify-content: center;
 }
+
+.spacing {
+  margin-bottom: 30px;
+}
