--- src/features/apiDelegation/components/SummaryPage/index.ts
@@ -1,1 +1,0 @@
-export { SummaryPage } from './SummaryPage';
